Remove non-core configuration parameters add custom config instead
This allows for the Probes to use probe specific configuration.  The
parameters are key value pairs that are stored in the database as
JSON.  The data is then mapped back to a Map<String, String> in
the java code.  A custom Converter is used to hide this from the
application code.