Re-factor Configuration mechanism
Each probe has it's own Configuration sub-class that is marked in the
ProbeMarker Annotation.  The model classes are single table inheritance
sub-classes.  The data extra data is stored in the map table and is
persisted as JSON in the database.  This should allow for a stronger typing
in the probe classes.