Yet another re-do of the configuration system
We are going back to the simple single class for a ProbeConfiguration.  The
complexity of determining the wrapper for the config, is now in the java
code itself, rather than relying on the hibernate to sort it out for us.

The ProbeMarker now specifies the type name of the ProbeType and the Class
that hold the probe specific configuration getters/setters.  An additional
piece of Reflection voodoo was added to instantiate the probe specific
config and seeding it with the probeMarkerCache

The unit tests needed to be re-visited for this to work.